change when the configuration was saved
Changing the relay switching sequence and saving the new configuration.
First, we save the new config, in case the ESP-8266 fails when switching the relay (EMP), this allows us to get rid of the possible destruction of the configuration file during recording.
The disadvantage is the delay in reaction from pressing the button.
If the commutation is weak or "correct" wiring, then storage and switching can be reversed.